fix for serialization issues

+++ src/main/java/org/jbei/ice/client/collection/presenter/CollectionsPresenter.java
@@ -1,5 +1,6 @@
 package org.jbei.ice.client.collection.presenter;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.Set;
@@ -418,7 +419,7 @@ public class CollectionsPresenter extends AbstractPresenter {
 
         // RPC with newName
         FolderDetails editFolder = new FolderDetails();
-        editFolder.setCount(item.getCount());
+        editFolder.setCount(BigInteger.valueOf(item.getCount()));
         editFolder.setName(newName);
         editFolder.setId(item.getId());
 
@@ -435,7 +436,7 @@ public class CollectionsPresenter extends AbstractPresenter {
 
                 display.updateSubMenuFolder(new OptionSelect(folder.getId(), folder.getName()));
                 MenuItem resultItem = new MenuItem(folder.getId(), folder.getName(), folder
-                        .getCount(), folder.isSystemFolder());
+                        .getCount().longValue(), folder.isSystemFolder());
                 display.setMenuItem(resultItem, deleteHandler);
             }
         });
@@ -458,8 +459,8 @@ public class CollectionsPresenter extends AbstractPresenter {
 
                 userListProvider.getList().add(folder);
                 display.addSubMenuFolder(new OptionSelect(folder.getId(), folder.getName()));
-                MenuItem newItem = new MenuItem(folder.getId(), folder.getName(),
-                        folder.getCount(), folder.isSystemFolder());
+                MenuItem newItem = new MenuItem(folder.getId(), folder.getName(), folder.getCount()
+                        .longValue(), folder.isSystemFolder());
 
                 if (!folder.isSystemFolder())
                     display.addMenuItem(newItem, deleteHandler);
@@ -563,8 +564,8 @@ public class CollectionsPresenter extends AbstractPresenter {
             ArrayList<FolderDetails> systemFolder = new ArrayList<FolderDetails>();
 
             for (FolderDetails folder : folders) {
-                MenuItem item = new MenuItem(folder.getId(), folder.getName(), folder.getCount(),
-                        folder.isSystemFolder());
+                MenuItem item = new MenuItem(folder.getId(), folder.getName(), folder.getCount()
+                        .longValue(), folder.isSystemFolder());
 
                 if (folder.isSystemFolder()) {
                     systemMenuItems.add(item);
@@ -644,7 +645,7 @@ public class CollectionsPresenter extends AbstractPresenter {
 
                         ArrayList<MenuItem> items = new ArrayList<MenuItem>();
                         MenuItem updateItem = new MenuItem(result.getId(), result.getName(), result
-                                .getCount(), result.isSystemFolder());
+                                .getCount().longValue(), result.isSystemFolder());
                         items.add(updateItem);
                         display.updateMenuItemCounts(items);
